+++ go/tasks/plugins/k8s/sagemaker/custom_training.go
@@ -140,7 +140,7 @@ func (m awsSagemakerPlugin) getTaskPhaseForCustomTrainingJob(
 	ctx context.Context, pluginContext k8s.PluginContext, trainingJob *trainingjobv1.TrainingJob) (pluginsCore.PhaseInfo, error) {
 
 	logger.Infof(ctx, "Getting task phase for sagemaker training job [%v]", trainingJob.Status.SageMakerTrainingJobName)
-	info, err := m.getEventInfoForJob(ctx, trainingJob)
+	info, err := m.getEventInfoForCustomTrainingJob(ctx, trainingJob)
 	if err != nil {
 		return pluginsCore.PhaseInfoUndefined, pluginErrors.Wrapf(pluginErrors.RuntimeFailure, err, "Failed to get event info for the job")
 	}
@@ -184,4 +184,18 @@ func (m awsSagemakerPlugin) getTaskPhaseForCustomTrainingJob(
 	}
 
 	return pluginsCore.PhaseInfoRunning(pluginsCore.DefaultPhaseVersion, info), nil
+}
+
+func (m awsSagemakerPlugin) getEventInfoForCustomTrainingJob(ctx context.Context, trainingJob *trainingjobv1.TrainingJob) (*pluginsCore.TaskInfo, error) {
+
+	var jobRegion, jobName, jobTypeInURL, sagemakerLinkName string
+	jobRegion = *trainingJob.Spec.Region
+	jobName = *trainingJob.Spec.TrainingJobName
+	jobTypeInURL = "jobs"
+	sagemakerLinkName = CustomTrainingJobSageMakerLinkName
+
+	logger.Infof(ctx, "Getting event information for SageMaker CustomTrainingJob task, job region: [%v], job name: [%v], "+
+		"job type in url: [%v], sagemaker link name: [%v]", jobRegion, jobName, jobTypeInURL, sagemakerLinkName)
+
+	return createTaskInfo(ctx, jobRegion, jobName, jobTypeInURL, sagemakerLinkName)
 }
