+++ go/tasks/plugins/k8s/sagemaker/custom_training_test.go
@@ -6,6 +6,8 @@ import (
 	"strconv"
 	"testing"
 
+	flyteIdlCore "github.com/lyft/flyteidl/gen/pb-go/flyteidl/core"
+
 	commonv1 "github.com/aws/amazon-sagemaker-operator-for-k8s/api/v1/common"
 	"github.com/lyft/flyteplugins/go/tasks/plugins/k8s/sagemaker/config"
 
@@ -15,6 +17,7 @@ import (
 	trainingjobv1 "github.com/aws/amazon-sagemaker-operator-for-k8s/api/v1/trainingjob"
 	"github.com/aws/aws-sdk-go/service/sagemaker"
 	pluginsCore "github.com/lyft/flyteplugins/go/tasks/pluginmachinery/core"
+	"github.com/lyft/flyteplugins/go/tasks/pluginmachinery/utils"
 	stdConfig "github.com/lyft/flytestdlib/config"
 	"github.com/lyft/flytestdlib/config/viper"
 )
@@ -134,5 +137,78 @@ func Test_awsSagemakerPlugin_GetTaskPhaseForCustomTrainingJob(t *testing.T) {
 		phaseInfo, err := awsSageMakerTrainingJobHandler.getTaskPhaseForCustomTrainingJob(ctx, taskCtx, trainingJob)
 		assert.NotNil(t, err)
 		assert.Equal(t, phaseInfo.Phase(), pluginsCore.PhaseUndefined)
+	})
+}
+
+func Test_awsSagemakerPlugin_getEventInfoForCustomTrainingJob(t *testing.T) {
+	// Default config does not contain a roleAnnotationKey -> expecting to get the role from default config
+	ctx := context.TODO()
+	defaultCfg := config.GetSagemakerConfig()
+	defer func() {
+		_ = config.SetSagemakerConfig(defaultCfg)
+	}()
+
+	t.Run("get event info should return correctly formatted log links for custom training job", func(t *testing.T) {
+		// Injecting a config which contains a mismatched roleAnnotationKey -> expecting to get the role from the config
+		configAccessor := viper.NewAccessor(stdConfig.Options{
+			StrictMode: true,
+			// Use a different
+			SearchPaths: []string{"testdata/config2.yaml"},
+		})
+
+		err := configAccessor.UpdateConfig(context.TODO())
+		assert.NoError(t, err)
+
+		awsSageMakerTrainingJobHandler := awsSagemakerPlugin{TaskType: customTrainingJobTaskType}
+
+		tjObj := generateMockTrainingJobCustomObj(
+			sagemakerIdl.InputMode_FILE, sagemakerIdl.AlgorithmName_CUSTOM, "", []*sagemakerIdl.MetricDefinition{},
+			sagemakerIdl.InputContentType_TEXT_CSV, 1, "ml.m4.xlarge", 25)
+		taskTemplate := generateMockTrainingJobTaskTemplate("the job", tjObj)
+		taskCtx := generateMockCustomTrainingJobTaskContext(taskTemplate, false)
+		trainingJobResource, err := awsSageMakerTrainingJobHandler.BuildResource(ctx, taskCtx)
+		assert.NoError(t, err)
+		assert.NotNil(t, trainingJobResource)
+
+		trainingJob, ok := trainingJobResource.(*trainingjobv1.TrainingJob)
+		assert.True(t, ok)
+
+		taskInfo, err := awsSageMakerTrainingJobHandler.getEventInfoForCustomTrainingJob(ctx, trainingJob)
+		if err != nil {
+			panic(err)
+		}
+
+		expectedTaskLogs := []*flyteIdlCore.TaskLog{
+			{
+				Uri: fmt.Sprintf("https://%s.console.aws.amazon.com/cloudwatch/home?region=%s#logStream:group=/aws/sagemaker/TrainingJobs;prefix=%s;streamFilter=typeLogStreamPrefix",
+					"us-west-2", "us-west-2", taskCtx.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName()),
+				Name:          CloudWatchLogLinkName,
+				MessageFormat: flyteIdlCore.TaskLog_JSON,
+			},
+			{
+				Uri: fmt.Sprintf("https://%s.console.aws.amazon.com/sagemaker/home?region=%s#/%s/%s",
+					"us-west-2", "us-west-2", "jobs", taskCtx.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName()),
+				Name:          CustomTrainingJobSageMakerLinkName,
+				MessageFormat: flyteIdlCore.TaskLog_UNKNOWN,
+			},
+		}
+
+		expectedCustomInfo, _ := utils.MarshalObjToStruct(map[string]string{})
+		assert.Equal(t,
+			func(tis []*flyteIdlCore.TaskLog) []flyteIdlCore.TaskLog {
+				ret := make([]flyteIdlCore.TaskLog, 0, len(tis))
+				for _, ti := range tis {
+					ret = append(ret, *ti)
+				}
+				return ret
+			}(expectedTaskLogs),
+			func(tis []*flyteIdlCore.TaskLog) []flyteIdlCore.TaskLog {
+				ret := make([]flyteIdlCore.TaskLog, 0, len(tis))
+				for _, ti := range tis {
+					ret = append(ret, *ti)
+				}
+				return ret
+			}(taskInfo.Logs))
+		assert.Equal(t, *expectedCustomInfo, *taskInfo.CustomInfo)
 	})
 }
